Move updatable macro types to /models
