+++ src/client.spec.ts
@@ -36,7 +36,6 @@ import {
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
 	TargetedNotificationTargetType,
-	UpdatableMacro,
 	UpdatableNotification,
 	UpdatablePlaybook,
 	UpdatableResource,
@@ -88,6 +87,7 @@ import {
 	UpdatableDashboard,
 	UpdatableFile,
 	UpdatableGroup,
+	UpdatableMacro,
 	UploadableAutoExtractor,
 	User,
 	UserPreferences,
