+++ src/functions/macros/update-one-macro.ts
@@ -6,9 +6,7 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { isUndefined } from 'lodash';
-import { Macro, RawMacro, toMacro } from '../../models';
-import { NumericID, RawNumericID, toRawNumericID } from '../../value-objects';
+import { Macro, RawMacro, toMacro, toRawUpdatableMacro, UpdatableMacro } from '../../models';
 import {
 	APIFunctionMakerOptions,
 	buildHTTPRequest,
@@ -44,34 +42,3 @@ export const makeUpdateOneMacro = (makerOptions: APIFunctionMakerOptions) => {
 		}
 	};
 };
-
-export interface UpdatableMacro {
-	id: NumericID;
-	groupIDs?: Array<NumericID>;
-
-	name?: string;
-	description?: string | null;
-	labels?: Array<string>;
-
-	expansion?: string;
-}
-
-interface RawUpdatableMacro {
-	GIDs: Array<RawNumericID>;
-
-	Name: string;
-	Description: string; // Use empty string for null
-	Labels: Array<string>;
-
-	Expansion: string;
-}
-
-const toRawUpdatableMacro = (updatable: UpdatableMacro, current: Macro): RawUpdatableMacro => ({
-	GIDs: (updatable.groupIDs ?? current.groupIDs).map(toRawNumericID),
-
-	Name: updatable.name ?? current.name,
-	Description: (isUndefined(updatable.description) ? current.description : updatable.description) ?? '',
-	Labels: updatable.labels ?? current.labels,
-
-	Expansion: updatable.expansion ?? current.expansion,
-});
